worker/pyproject.toml
[project]
name = "mlx-worker"
version = "0.1.0"
description = "mlx-lm grpc worker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.78.0",
    "mlx-lm>=0.30.6",
]
